backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
alembic>=1.12.1
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
open_clip_torch>=2.20.0
torch>=2.0.0
pillow>=10.0.0
